+++ src/hedged_file_system.cpp
@@ -369,8 +369,7 @@ FileMetadata HedgedFileSystem::Stats(FileHandle &handle) {
 // HedgedFileHandle
 //===--------------------------------------------------------------------===//
 
-HedgedFileHandle::HedgedFileHandle(HedgedFileSystem &fs, unique_ptr<FileHandle> wrapped_handle,
-                                   const string &path)
+HedgedFileHandle::HedgedFileHandle(HedgedFileSystem &fs, unique_ptr<FileHandle> wrapped_handle, const string &path)
     : FileHandle(fs, path, wrapped_handle->GetFlags()), hedged_fs(fs),
       wrapped_handle(wrapped_handle.release(), [](FileHandle *handle) {
 	      if (handle) {