+++ CHANGELOG.md
@@ -3,3 +3,7 @@
 ### Changed
 
 - Update DuckDB and extension-ci-tools to v1.5.0
+
+### Added
+
+- Add extension description